Remove get_contract_class implementation from trait
